Ingrese en el Fix master y arregle todo

+++ src/main/java/com/progra3/proyecto/repository/VehiculoRepository.java
@@ -7,7 +7,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.progra3.proyecto.entity.Vehiculo;
 
-public interface VehiculoRepository extends JpaRepository<Vehiculo, Integer> {
+public interface VehiculoRepository extends JpaRepository<Vehiculo, Long> {
     void deleteById(Long id);
     boolean existsById(Long id);
     List<Vehiculo> findAll();
